KVM: vmx/pmu: Create a guest LBR event when vcpu sets
From 8e12911b243e485f5e4c7c5fbc79cdf185728700 Mon Sep 17 00:00:00 2001
 DEBUGCTLMSR_LBR
Git-commit: 8e12911b243e485f5e4c7c5fbc79cdf185728700
Patch-mainline: v5.12-rc2
References: bsn#12

When vcpu sets DEBUGCTLMSR_LBR in the MSR_IA32_DEBUGCTLMSR, the KVM handler
would create a guest LBR event which enables the callstack mode and none of
hardware counter is assigned. The host perf would schedule and enable this
event as usual but in an exclusive way.

The guest LBR event will be released when the vPMU is reset but soon,
the lazy release mechanism would be applied to this event like a vPMC.

Suggested-by: Andi Kleen <ak@linux.intel.com>
Co-developed-by: Wei Wang <wei.w.wang@intel.com>
Signed-off-by: Wei Wang <wei.w.wang@intel.com>
Signed-off-by: Like Xu <like.xu@linux.intel.com>
Reviewed-by: Andi Kleen <ak@linux.intel.com>
Message-Id: <20210201051039.255478-6-like.xu@linux.intel.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.mainline/v5.12-rc2-KVM-vmx-pmu-Create-a-guest-LBR-event-when-vcpu-sets-.patch