mptcp: create the listening socket for new port
Git-commit: 1729cf186d8a5d70cf7a54e07c4763635079f015
Message-Id: <1729cf186d8a5d70cf7a54e07c4763635079f015.1637074505.git.geliang.tang@suse.com>
References: bsn#2
Patch-mainline: v5.12-rc1

This patch creates a listening socket when an address with a port-number
is added by PM netlink. Then binds the new port to the socket, and
listens for new connections.

When the address is removed or the addresses are flushed by PM netlink,
release the listening socket.

Signed-off-by: Geliang Tang <geliangtang@gmail.com>
Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>

Patch-name: patches.suse/v5.12-rc1-mptcp-create-the-listening-socket-for-new-port.patch